kernel: fix regression in max310x/regmap/regcache
The kernel driver for our UART was buggy. The data that were written out
to the Nistica WSS were ending up in the kernel's regmap's regcache
layer, and as a result, the UART traffic was only sent the first time.
Subsequent opens of the port would end up in a total silence.

This turned out to be fixed independently on mainline, so just upgrade
to v6.6.3 to get that fix. The current branch is available form our
Gerrit copy of the kernel, branch `cesnet/2023-12-02`. I also used this
opportunity to bring in some max310x patches there were posted for
review -- see the changelog in the kernel repo if you're interested.
Here's what we're carrying now:

Hugo Villeneuve (2):
      serial: max310x: add macro for max number of ports
      serial: max310x: use separate regmap name for each port

Jan Kundrát (6):
      spi: orion: Respect per-transfer bits_per_word settings
      ARM: dts: armada-388-clearfog: support I2C recovery on i2c-1
      pinctrl: mcp23s08: work around GPIO line naming
      leds: tlc591xx: SW reset during initialization
      Revert "leds: Mark GPIO LED trigger broken"
      tty: max310x: Fail probe when external clock crystal is not stable

Bug: https://lore.kernel.org/linux-serial/bd91db46c50615bc1d1d62beb659fa7f62386446.1701446070.git.jan.kundrat@cesnet.cz/T/#u
Change-Id: Ibfcdfb995eeae6a60b81f313513194ca20220c18