Sync velia
It seems one of our boxes started behaving crazy. It reports 2.3kW
input power in one of the PSUs while usually the values are around 21W.
The value is way beyond what ietf-hardware YANG model can represent and
libyang throws a validation error which kills velia and our whole NETCONF
stack.
Logs attached below (thanks to Jan Kundrát for those).

This brings in the patch that tries to address the issue when trying to
write invalid value into sysrepo. It *does not* fix the HW issue but
rather it logs an error and signalizes an overflow/underflow in the
sensor data.

Some logs:

 * sysfs reads of the power_input values: Note the values at
14:11:29:

  add-drop-DQ000VOT ~ # while true; do date; cat /sys/class/hwmon/hwmon8/power*_input ; sleep 1; done
  ...
  Mon Dec 11 14:11:27 UTC 2023
  25250000
  21000000
  Mon Dec 11 14:11:28 UTC 2023
  25500000
  21000000
  Mon Dec 11 14:11:29 UTC 2023
  2316000000
  20000000
  Mon Dec 11 14:11:30 UTC 2023
  25250000
  21000000
  ...
  Mon Dec 11 14:12:01 UTC 2023
  25250000
  20000000
  Mon Dec 11 14:12:02 UTC 2023
  0
  20000000
  Mon Dec 11 14:12:03 UTC 2023
  25000000
  20000000
  ...

 * and the original velia crash

  Dec 11 13:59:46 add-drop-DQ000VOT veliad-hardware[7997]: terminate called after throwing an instance of 'libyang::ErrorWithCode'
  Dec 11 13:59:46 add-drop-DQ000VOT veliad-hardware[7997]:   what():  Couldn't create a node with path '/ietf-hardware:hardware/component[name='ne:psu2:power-in']/sensor-data/value': LY_EVALID
  ...
  Dec 11 13:59:46 add-drop-DQ000VOT main[7997]: Processing node update /ietf-hardware:hardware/component[name='ne:psu2:power-in']/sensor-data/value -> 2316000000
  Dec 11 13:59:47 add-drop-DQ000VOT systemd[1]: velia-hardware-g2.service: Main process exited, code=dumped, status=6/ABRT
  Dec 11 13:59:47 add-drop-DQ000VOT systemd[1]: velia-hardware-g2.service: Failed with result 'core-dump'.

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/velia/+/6703
Change-Id: I07ef810a69842e3e37910ae3a427eaf432e1c00e