Sync velia
It seems one of our boxes started behaving crazy. It reports 39kW
input power in one of the PSUs.
The value is way beyond what ietf-hardware YANG model can represent and
libyang throws a validation error which kills velia and our whole NETCONF
stack.

This brings in the patch that tries to address the issue when trying to
write invalid value into sysrepo. *It does not fix the HW issue.*

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/velia/+/6703
Change-Id: I07ef810a69842e3e37910ae3a427eaf432e1c00e