kernel: report stuck I2C bus during bootup
We're using this GPIO-based, bit-banged I2C bus for a PMBus connection
towards the PSUs. Every now and then, we're getting some weird errors
related to the PMBus implementation, such as the device reporting utter
garbage for some hwmon/pmbus attribute, or a sysfs file not being
readable or not containing an integer.

Sometimes, though, the bus appears to be really stuck, with the most
recent occurrence manifesting as SDA being high (which is good, when
it's idle) and SCL stuck low, as shown in `/sys/kernel/debug/gpio`:

 gpio-24  (scl                 ) in  lo (act hi) - IRQ disabled
 gpio-25  (sda                 ) in  hi (act hi) - IRQ disabled

This is very very bad, because a stuck SCL means that we cannot do
absolutely anything; even the classic "stuck I2C recovery" of repeatedly
driving SCL low cannot be used (because it's already stuck low). Since
we cannot do anything to fix this, the least that we *can* do is
to report this condition.

We, however, still register the entire I2C bus. The individual I2C
slaves will fail to get probed, that's just tough life, but at least the
bus itself is "available". In the only case we've hit this, so far, the
bug was caused by a PSU module that can be hot-unplugged. Theoretically,
with proper support within velia, which already has dynamic registration
for the PSUs, we can extend this for the PDU as well, and have full
hot-pluggability.

TL;DR: at least issue that warning during bootup.

Change-Id: I422be78ec9001c734c49634190e95e81d4b6f350