Sync velia
There are two new features:

- remote log upload
- PDU "hot-plugging"

The log upload requires the usual separation between startup and
running, so there's now a second EnvironmentFile for the service.

The PDU "hot-plugging" is a workaround for some device-specific errors
caused by the PSUs (which share the I2C bus with the PDU). Apparently,
the bus would end up with the SCL held low for hours/days/indefinitely,
and we have to keep velia-health running even if that is the case. The
root cause, which is "something wrong" in the PMBus communication, is
not fixed, but at least the whole system won't get unusable.
Unfortunately, a proper alarm is missing so far.

Change-Id: I72c1c5249077b99b3858dcb956b53e97b5505f60