velia: fix persistent location of /etc/hostname
The service velia-system has two subscriptions within sysrepo, one for
the startup DS, and the other for running. The subscription which is
responsible for the startup DS is supposed to "pregenerate" the
/etc/hostname for the next reboot, which is OK, except that on our
system the /etc is an overlay filesystem, and therefore the change
wasn't taking any effect, and was only applied when the velia-system
started and applied its running state.

Fix that by storing the persistent hostaname below /cfg/etc which is
restored through the RAUC hook (see
package/czechlight-rauc/rauc-hook.sh).

Change-Id: Ia198528a5f54a8bb712c6293ba33cdb4ca82d04c