sysrepo: ensure that stuck /run/sysrepo SHM can be trashed
When the sysrepo stack dies, we were trying hard to just shoot all the
services which use that filesystem to restart, unmount that, and start
from scratch. Unfortunately, this was not really happening. When there
was any lingering process that was still accessing the filesystem, such
as `sysrepo-cli`, the unmount would fail.

Fix that by using `-f` (which I was thinking would do the job, but nope,
it doesn't), and `-l`. Together, they do the trick.

Change-Id: Icb0c706d8334f8434df0e4b97cccca893a2c9839