Upgrade buildroot
- buildroot is now at `cesnet/2024-01-16~1`
- systemd patches are from Tomas' `lldp-json-254.6` branch

Unfortunately, upstream Buildroot now has a systemwide mitigation
against using `.git` and/or the `.gitdir` mechanism in there for
overriding projects' git version. That's a bummer because we're using
just that within the CI to preserve version information, as provided by
Zuul. Upstream apparently wants to keep that, so let's make sure that we
reset the GIT_ENV at the build time for our packages. That's better than
doing a revert because other packages in Buildroot might rely on this.
Well, we have a check in `dev-setup-git.sh` already for not building
inside the br2-external repo, but hey, let's be fine-grained here.

Change-Id: I11869f981817e4c2c8322c14cde8cadb49cb9927
Bug: https://lore.kernel.org/buildroot/20240111105621.2321957-1-peter@korsgaard.com/T/#mb1ce8cebca272ab94c47b56e58232a59e1af1e94