Upgrade buildroot
- buildroot is now at `cesnet/2024-01-16`
- systemd patches are from Tomas' `lldp-json-254.6` branch

Unfortunately, upstream Buildroot now has a systemwide mitigation
against using `.git` and/or the `.gitdir` mechanism in there for
overriding projects' git version. That's a bummer because we're using
just that within the CI to preserve version information, as provided by
Zuul. As a temporary workaround before it gets solved upstream, let's
revert that upstream commit which introduced this change.

I'm not super happy with that because it was a part of a series which
removed some GIT_DIR overrides in other projects, but OTOH we never
build under a git-managed directory for exactly this reason, and there's
an explicit check for this in `dev-setup-git.sh`, so we should not be
affected. At least until someone inits a git repo among some of the
parents of the build dir after it's been set up, or something silly like
that.

Change-Id: I11869f981817e4c2c8322c14cde8cadb49cb9927
Bug: https://lore.kernel.org/buildroot/20240111105621.2321957-1-peter@korsgaard.com/T/#mb1ce8cebca272ab94c47b56e58232a59e1af1e94