Update buildroot
This is your regular sync to master + rebase of our local patches, plus
fixes of whatever got broken. Commits are available in the
`cesnet/2024-04-22` branch.

One semi-notable thing is a fixup for the libnetconf2 build because of
the usual `-lcrypt` transition; it used to be provided by glibc, but it
ain't no more.

Also, Ruby got upgraded which needed some tweaks to make sure that mruby
still worked in nghttp2-apps' nghttpx.

Change-Id: Ibf3edf0932418dfef41e4cf283c41e2be1906c48