migrations: access correct project repositories
The test script assumed that the real project's source code was always
saved in a git submodule. That's not the case when Zuul speculatively
merges a change ahead, and it's also not necessarily true on each
developer's machine (it surely wasn't for me).

Fix that by moving the actual invocation into make. However, since
Buildroot's per-package directories are a thing, this one really needs a
BR_PATH that's expanded to include the "current package", which means
that we have to tweak some extra variable names. For regular builds,
these are done via packages/pkg-generic.mk and the generic-package
macro, but we have to define PKG and the expanded $(PKG)_NAME, otherwise
the test suite won't find, say, `sysrepoctl`.

The actual migrations refer to some YANG models and maybe also YANG data
from additional packages. To make these available in a place where
Buildroot expects them, one could either depend on, e.g., files such as
`velia_CONFIGURE`, or we could just as well add them all into the
dependency list of this package, and then depend on this package's own
`configure` step to have finished.

I guess I don't have to say it loud, but hey, this has been one hell of
a ride to get done correctly. The alternative would be to extract the
<PKG>_OVERRIDE_SRCDIR from Make (`make printvars
VARS=NETOPEER2_OVERRIDE_SRCDIR`), but I felt like this one was "better
integrated".

Change-Id: I2887482f0aa788bedf37548d8ba5010b54c29679