tests: migrations: purge state in between runs
These migrations were not idempotent because of sysrepo's state that's
preserved in SHM. Make sure that this is thrown away properly before
each test run. The easiest fix is via running these below a shared
prefix.

Previously, there would be a unique prefix for each migration run.
That's fine. We *could* also use a common prefix which is the same for
all migrations with this patch, but let's allow us some space for a
possible future parallelization here.

Change-Id: I33c2e093dbc375b2dbdeaae02aa430ee6f486ae3