ci: show details why tree prepared by Zuul differs from the submodules/PROJECT tree
I was dealing with some upgrades in br2-external submodules and I was
confused because the log said that the one of the project's submodules
git tree was different than expected. The commit hash was also not in my
local git tree. I was confused.
It turns out that sometimes Zuul can create merge commits.
Let's print the last 5 commits from the tree prepared by Zuul and the
last 5 commits from the tree in submodules/$project. I think that would
help greenhorns (or just me) who don't understand the ci internals.

Change-Id: I5e1f1ecd45415c51390de2b4eaef6db4ba866c9d