unbreak the USB flasher
It seems that somewhere during the past three (or so) years, systemd has
changed its behavior when it comes to how udev is managed. As a result,
the USB flasher which was set to run "without extra dependencies" did
not work:

 systemd[1]: dev-sda1.device: Job dev-sda1.device/start timed out.

...while /dev/sda1 device was obviously detected by the kernel. Rather
than trying to figure out what exactly is broken, let's convert this to
a "regular" systemd service without the `DefaultDependencies=no` stanza.
This required changing the watchdog activation since systemd happily
takes that one over.

Also, since commit 9dec725 ("Sync all projects"), we now have
`/usr/sbin/blkdiscard` from util-linux and the `/usr/bin/blkdiscard`
from busybox. The former requires the `-f` flag to force it to do stuff,
and the latter's `--help` doesn't say anything about that -- but hey, at
least it (silenty) skips it.

Change-Id: I188fffc231820376bcf2d92404092427091a5697