Upgrade U-Boot to the latest upstream
...which is v2024.10. This required some changes:

- Since upstream commit 87ac4b4b4ca5f00e2ddcdac41c9dc691ab2aecf1 (), the
  name of the KWB image has changed. This propagates all over the place,
  to various flashing scripts, etc.

- I moved the redundant env configuration to a config file so that we
  carry fewer patches.

- Our fork is available on Gerrit under the cesnet/2024-10-11 branch. We
  have just two patches now:

  - a2db423a26a clearfog: ext board: LED blinking at power-up
  - 082f86a75b2 clearfog: custom config for redundant boot via RAUC

  Since these are reasonable self-contained now, we can switch to using
  a release tarball instead of a git snapshot.

- The other config change is just a small fixup to not depend on
  libuuid/util-linux since this is not an EFI platform, anyway (see the
  KConfig README in buildroot's u-boot packaging for details).

Unfortunately, this new .kwb file apparently has a slightly different
format from the previous one, and as such, `kwbtool` from
uboot-tools-2021.7 (which is what is even in the *latest* Buildroot)
cannot work with this format:

 Sending boot image...
 0 % [+xmodem: Protocol error

So, let's upgrade uboot-tools, right? Hell no -- that thing is stuck on
2021.7 for a reason, there's some dark voodoo for ignoring the board
config, and I *cannot* port that to 2024.10 (and I really tried hard).
Besides, Buildroot by default doesn't include `kwboot` when building
uboot-tools, that was always done by a patch of mine.

So, rather than trying to upgrade uboot-tools and to reverse engineer
all that logic for configuring of that package, let's give up and build
a new enough `kwboot` from scratch. That was far from trivial (see the
new `kwboot.mk` in this repo), and since this thing now depends on
linking with ncurses' `-ltinfo`, which is obviously *not* provided on
Buildroot by that package, let's stop building `kwboot` for the target,
and built that thing only on host. As a result, one cannot use a
Clearfog board to bootstrap another Clearfog board anymore. I can live
with that.

The "updated" Buildroot at the `cesnet/2024-10-11` branch is simply a
rebase with a single patch dropped:

 - 6068afca3b uboot-tools: Allow building `kwboot`

Change-Id: I07858c7711ee31e8dc6ac2e07b7ea73913a14ca0