pmbus: reduce wait time before we declare the I2C bus stuck
By default, the bitbang driver waits up to 100ms for the SCL to raise
back up (and it has been doing that since the initial upstream commit
back in 2007). That's way too long, and the driver will wait this long
for the first clock cycle *and* for the STOP condition as well. That's
200ms so far, and then the kernel happily retries stuff again.

These retries are done 3 times, and this arbitrary number goes beyond
the git and BitKeeper histories, combined. I bumped upon some "tweaks"
to the number of retries [1] from 2003 which mention the lm_sensors
repo, so I suspect that someone used a really crappy jumper wire back in
the 90s, and since that time it's been a rule that A Failing Transaction
Must Be Retried, Sometimes (but only on a bit-banged I2C, of course).

Let's contrast this with the SMBus spec which is perfectly clear ([2],
Table 2, pages 18 and 19, parameter t_TIMEOUT) that any device can
detect a bus as stuck after 25ms, and must do so within 35ms.

-> Let's set the timeout to 25ms which is enough™ for any compilant
device.

[1] https://github.com/jeffmahoney/linux-pre-git/commit/88ede345d538dcf073f91beecd11a5c3425e194c
[2] https://www.smbus.org/specs/SMBus_3_3_1_20241020.pdf

Change-Id: I4b146041113473cf704b12d540d7245e38a35432