clearfog: build a preemptible kernel
...because I really want to cut these 200ms+ latencies (as reported by
`rtla timerlat`) down a *lot*, and because the description of this
config option sounds like something that I absolutely want to have.
Also, it's one step before going the full PREEMPT_RT route.

On a box where the I2C-2 bus (the bit-banged one) is locked up, I used
to see latencies of about 200ms -- that was prior to the recent
optimizations of I2C retries in [1] and [2]. With these two patches, it
went down to about 80ms. With this one, I'm at less than 0.5ms, and
that's with a running `perf`.

(The other difference is that the cla-sysrepod is not crashing with UART
RX FIFO overruns. This is the desired outcome, but it "might" affect the
performance measurements because there are no extra instances of the
systemd-coredump@.service, which means that there are fewer updates to
the ietf-alarms' operational DS, etc. Did I mention that this is a
moving target where everything affects everything else?)

[1] https://gerrit.cesnet.cz/c/CzechLight/br2-external/+/7879
[2] https://gerrit.cesnet.cz/c/CzechLight/velia/+/7884

Change-Id: I72b2857288dcf370d03161a8a58985893f17c4d6