clearfog: faster bit-banged I2C bus access
There was a typo in DT property name, and therefore this option has
always been a no-op, since the very beginning. That means that the
kernel's default of 5µs has always been in effect.

I made some measurements, and it seems that using a custom delay of 3µs
does the trick here. With this option, the frequency stays at 100kHz
(and safely below 100kHz, at about 92-96kHz when loaded just a little
bit).

Fixes: 7f2ff837 clearfog: Add a SW bit-bang I2C bus for PMBus
Change-Id: I8bff086330aa8d8a15b4ba8bd2f5c8e261e4fe59