Implement additional PMBus quirks in the kernel
This is the `cesnet/2024-10-25--01` branch. The TL;DR version is that
the it *seems* that VOUT_MODE is not implemented on many more devices,
and when that's the case, one is expected to fall back to the linear11
encoding (which is not allowed for output voltages, but hey, who cares
about standards).

This is not just on some older modules, but also on the most recent DC
one from 2023.

Also, page updates were sometimes ignored (the device would return, say,
a 5V value on a 12V page, and the other way round). Let's increase the
timeout to 50ms and hope for the best^H^H^Hkeep an eye on how this thing
behaves now.

Change-Id: I56e8cbf660306332ef0fe5c23a75528d8eeec6f4