unbreak USB flashing, take two
There's a memory limit of "half the RAM" which applies to the sum of
sizes of the compressed and uncompressed CPIO archive of the rootfs. We
have been working dangerously close to this limit, and had to enable a
different compression recently just to buy us some time. This time has
run out recently, and now the board won't boot again, as reported by
Jarek.

We could have solved this by building a second standalone, minimal
rootfs, but I was too lazy to do that. Instead, simply filter the list
of files in the rootfs, selectively excluding the largest heavies from
the resulting CPIO image.

The change for Buildroot is available in our branch `cesnet/2024-11-15`.

Fixes: d08ced2a unbreak booting within the flash-via-USB image
Change-Id: If3962d80303cd5ddca6081e01e3054e97865cede