fix LED labels on various types of HW
This is a bit messy, I'm afraid. The only part which is common across
all device types is the PCB. According to the CAD data, its labeling
uses both D1/D2/... *and* the friendly labels such as STATUS and UID.
These friendly labels do not match labels on the case in some cases,
though. The silkscreen on the metal enclosure "almost always" has a
human-readable label which matches the LED's function. These functions
are different over different HW models, and as a not-so-nice bonus, some
inline amps use different labeling schemes, maybe due to some HW reuse
in the past.

So, let's setup the common parts in the "inner" DTSI file, *and* use
some DT-level labels to enable its overriding in the "outer" DT files.
Once these labels (which are visible in sysfs as well) are set up, let's
configure the operation from the userspace init script which sets up
proper kernel-level LED triggers.

Change-Id: I16f598e8e0b48e0fa71c35440645efedca7862f1