Support the new DC PDU
Thew new FSP PDUs for the DC pluggables have a different part number
which propagated all the way to the PMBUS_MFR_MODEL. As such, this
required a slight touch to the kernel, and to make stuff more
future-proof, also a small fix to velia.

The new kernel is in the `cesnet/2025-03-06` branch.

Change-Id: I475d3c333d201fd690433f6cc292750b5156d906
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/velia/+/8367