Fix LEDs on bidi amplifiers
The EDFA modules indicate loss-of-signal in the west-to-east direction
via the "stage 1 input LOS alarm", and on the east-to-west direction via
the "stage 2 input LOS alarm". As a result, we have two separate signals
that we can use to drive the LED status, so let's map the west input to
the green LED and the east input to the blue LED. When both are up, the
final result is a cyan light, and it looks almost white-ish to my tired
eye. Now, this is obviously not the best color coding, one might
conceivably want "green" to mean "all OK", but we cannot go that far
with a kernel-based GPIO-to-LED mapping, and I'm too lazy to write a
userspace GPIO-to-LED daemon tonight.

I might need to write one, though, because it turned out that these bidi
EDFAs use a different polarity on their output alarm pins. The datasheet
has always said that these are "active high", but nope, in the past, on
all the dual-stage fixed-gain extended C-band EDFAs that I've ever seen,
it's been active low. These modules have "version 1.0.1", whatever that
is.

On the new bidi EDFAs (both C-band and the 1572nm L-band), these pins
appear to be truly active-high. This seems to apply to all the stage1/2
input LOS, stage 1/2 out/gain, and pump current, case temperature, pump
temperature alarms.

I've sent a support question to the vendor. If this is really
FW-dependant, we *might* have to have a userspace thingy for this, maybe
in cla-sysrepo because that's the part which talks to the optical
modules already.

Change-Id: Ifadf16c024fd5d130ef7b145d0223a8676c9f642