LEDs: fix crash on inline amp
The first prototype of an inline amp was using some repurposed LEDs: the
"line" one became the "line east", and the original UID LED was called
"line west". (The front panel only had four holes, which were labeled
"STATUS", "WEST LINE", "EAST LINE" and "WEST OSC".) The LED init script
would properly set up the GPIO LED triggers, but there was still that
code in velia-system which expected a LED called "uid:<color>".
Everything worked "fine" up to the point where someone would try to
drive the UID lED, which would "also work", but afterwards the
GPIO-based indication would not really work anymore.

Then stuff changed in one of the previous commits where I introduced
device-dependant names for the sysfs LED nodes. Now, the LED bels made
sense, but since there was no sysfs entry called "uid:blue",
velia-system would crash with a not-so-great error message:

 Jan 01 00:00:09 EDFA-LINE-SDN-ID210512 systemd[1]: Started System management via sysrepo.
 Jan 01 00:00:11 EDFA-LINE-SDN-ID210512 veliad-system[267]: terminate called without an active exception
 Jan 01 00:00:14 EDFA-LINE-SDN-ID210512 systemd[1]: velia-system.service: Main process exited, code=dumped, status=6/ABRT
 Jan 01 00:00:14 EDFA-LINE-SDN-ID210512 systemd[1]: velia-system.service: Failed with result 'core-dump'.

Fortunately, the backtrace was kinda usable:

 Thread 1 (Thread 0xb667f020 (LWP 1143)):
 #0  0xb677bcc4 in ?? () from /lib/libc.so.6
 #1  0xb6732df4 in raise () from /lib/libc.so.6
 #2  0xb671e180 in abort () from /lib/libc.so.6
 #3  0xb696590c in __gnu_cxx::__verbose_terminate_handler() () from /lib/libstdc++.so.6
 #4  0xb6963484 in ?? () from /lib/libstdc++.so.6
 #5  0xb6958240 in std::terminate() () from /lib/libstdc++.so.6
 #6  0x00514a3c in std::__terminate () at /home/jkt/work/prog/_build/_br-cfb/per-package/velia/host/opt/ext-toolchain/arm-buildroot-linux-gnueabihf/include/c++/14.2.0/arm-buildroot-linux-gnueabihf/bits/c++config.h:324
 #7  std::thread::operator= (this=<optimized out>, __t=...) at /home/jkt/work/prog/_build/_br-cfb/per-package/velia/host/opt/ext-toolchain/arm-buildroot-linux-gnueabihf/include/c++/14.2.0/bits/std_thread.h:195
 #8  velia::system::LED::LED (this=0x81b0, this@entry=0xbef2380c, srConn=..., sysfsLeds=...) at /home/jkt/work/prog/_build/_br-cfb/build/velia-custom/src/system/LED.cpp:48
 #9  0x004a1730 in main (argc=<optimized out>, argv=<optimized out>) at /home/jkt/work/prog/_build/_br-cfb/build/velia-custom/src/main-system.cpp:121

Let's fix this by assigning two separate LEDs for the "signal present at
LINE XXX IN" port. The downside is that the first-ever prototype (with
no hole for led5) would have wrong labels. On the later prototypes with
holes for five LEDs, these would typically have labels of (<none>,
"STATUS", "UID", "LINE" and "SFP"). Since the "SFP" is already
kinda-misleading as the HW has *three* SFP cages, the built-in in the
CFB board which is labeled  "OSC WEST", an "OSC EAST" in the left cage
of the add-on PCIe card, and finally an "SFP" at the right slot of the
add-on card. The "SFP" LED however maps to the "OSC WEST" slot, so let's
just call it a day and give up trying to assign consistent labels to
anything. A label printer will solve this for me, I hope.

I'm not squashing this to
https://gerrit.cesnet.cz/c/CzechLight/br2-external/+/8452 because I feel
like the original problem has been around for a while.

Change-Id: Ie7c49a3d6ae88c34c768d56780688e47dfd3c0f6
Fixes: f974c910 fix LED labels on various types of HW