kernel: use batched reads on MAX14830, they reduce the CPU usage
I don't understand what I'm seeing -- we've got five pieces of these
bidi amps stacked on top of each other. They are "almost identical", the
only variation which we're getting is in the PDU/PSU modules. One of
them as an AC PDU (with a pair of AC, DC PSUs), the second one has a DC
PDU (with a pair of AC, DC PSUs), then three more devices with AC
PDU/PSUs. Apart from that, they are 100% identical as far as I can tell.

On #1, #3 and #5, I'm seeing about 20% CPU load from the soft IRQ
kthread which deals with spi1.2 (that would be the MAX14830 quad UART).
On #2 and #4, I see about 10% CPU load from that kernel thread. After
applying this patch, the CPU usage from that thread went to "almost
nothing" on host #2, but it remains at ~20% on host #1. Whiskey, Tango,
Foxtrot. I guess it's some profiling time ahead.

Change-Id: I08dff637a59105e3b73d4ce2f2e9de4563f1e786