doc: structure of the internal EEPROM
There's a standard ONIE TLV for "Device version" (0x26). I was thinking
whether it makes sense to use that instead of hardcoding the "-g2"
suffix in the "Part Number" (0x22), but in the end I decided to treat
the "new" version of the HW as an entirely new product. In the past,
we've done some iterating on the design, and the g1/g2 is mainly about
the differences on the main PCB, with a different electrical interface
towards the fan control board and a different I2C topology. Given that
the g1 boards are no longer used, and that we stopped caring about them
at all, and that the rest of the code "mostly" treats the entire
"sdn-foo-gX" as a single blob, I would prefer to stick to that naming
convention. Also, having yet another version number potentially allows
the vendor/manufacturer/... to iterate the design at least partially
independently from us. I realize that this is all very vague, but in the
end it "doesn't matter" whether we use two fields or a single field, so
a single field it is. Thanks to Tomas for suggestion the device version
approach.

Change-Id: Ib624fb4c1e63cc9b668f465c23547b2fa1a95556