sysrepo: reduce writes to disk when persisting the startup DS
To save some writes and make the thing more robust, let's check whether
the new file content is actually different, and only do the write when
there's been some modification.

We have to make the first write hit a tmpfs so that we actually save
some eMMC writes. This means that (if there are some changes) the final
destination is on the separate filesystem than the file that we just
wrote to via `sysrepocfg`. Now, there's no cross-filesystem rename(2),
and when I mv(1) these files, the `mv` actually truncates-and-overwrites
the destination file in-place, like this:

 openat(AT_FDCWD, "/cfg/xxx", O_WRONLY|O_CREAT|O_EXCL|O_LARGEFILE, 0100644)

This means that the target cannot be our final
/cfg/sysrepo/startup.json, and we have to copy through a temporary file
within /cfg (which we were trying to avoid writing to in case of no
differences in content). Hence we now do two `mv`s instead of just one
as we were doing previously. This is required to have a consistent
on-disk state upon a random reboot, or a power cut, etc.

The `PrivateTmp=yes` is to help hide this temporary file from the rest
of the system.

And while we're at this, let's use jq(1) for some pretty-formatting.
That should also make the diff (which is optional and commented out)
more readable.

Change-Id: I27c0bc8014981795d950bd4a881bcb4b998a4b4a