sysrepo: fix persisting the startup DS
Since 2017, we've been running a piece of code which attempted to detect
modifications to sysrepo's startup datastore, and when that happened, it
would dump the configuration into a persistent location on a writable
filesystem.

This might have worked at the beginning (or not). Over the time, we
switched from copying files around to using `sysrepocfg`, and sysrepo
might have started doing the write safely (or maybe it did so from the
very beginning). Anyway, the end result is that our code is broken
today:

- it only listens for writes to files, not necessarily to file moves,
- whenever an event happens, there's a period of no monitoring while we
  do the dump.

That's a race, and we've actually seen stuff not being persisted in the
past few days. Let's fix this by listening for a "file has been moved
here" event, and react to these events by dumping the DB content.

Change-Id: I4eddfaa058b4a4c6f4145c10ca94578fc4a4196d
Fixes: 41d2d267 ("Save persistent DSes to /cfg upon each change")
Fixes: 34d2f481 ("update and change sysrepo persistency")