kernel: unbreak access to igb's I2C bus, and update
As usual, this is from the `cesnet/2025-06-04` branch.

Unfortunately, this is "not enough" to get a simple hwmon-based access
to the SFP's TX/RX power. On Clearfog, these are provided by the `sfp`
driver, but the generic SFP layer is actually not used by igb at all.
Also, in the end we will very likely require reimplementing the readout
of these optical parameters in userspace because the hwmon abstractions
work in µW, and therefore their resolution is way too coarse (because
1 µW == -30dBm, 2 µW == -27 dBm, etc). Also, some SFPs have calibration
data, that might require FP math, and there's no FP in the kernel.

Change-Id: Id1bc237726e2911ebbd3c8cf2708cc2280ed4000