Update to libyang v4
Buildroot changes are in the `cesnet/2025-11-05` branch.

The local filesystem path of the netopeer2-server is now configurable
via YANG, and as a result, a malicious NACM-level administrator will be
able to write to the local filesystem as root. That's not good at all,
but upstream likes this design, and we have to mitigate that risk.
Setting a `must` which limits the actual path should be enough to
prevent permission escalation.

Also ensure that the Unix socket's endpoint cannot be accidentally
removed, or its permissions changed. OTOH, do not prescribe any uid/gid
changes, "just" prevent these leafs from being set to anything. That
way, we might switch to non-root at some other point without yet another
migration.

Change-Id: I85750267fdcf67b09b5358482f5a6ee6eafad9b8
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/8990
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/9001
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/netconf-cli/+/9000
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/rousette/+/8993